pyproject.toml
[project]
name = "text-sql-docker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arabic-reshaper>=3.0.0",
    "fastapi>=0.128.0",
    "ipykernel>=7.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "oracledb>=3.4.1",
    "pandas>=2.3.3",
    "python-bidi>=0.6.7",
    "regex>=2025.11.3",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.52.2",
    "uvicorn>=0.40.0",
]
